dm cache: write policy name to metadata in set_cache_policy
The policy name must be written to the metadata as soon as it is made
active.  Otherwise, subsequent table loads wouldn't know the currently
active policy and the code would treat said table load as if were the
first -- could've gotten the name from cache->policy but that wouldn't
deal with the case of changing the policy on first table load
(e.g. after a reboot).

I cannot see why waiting until suspend/destroy to write the policy name
is worthwhile anyway.