dm thin: replace calls to dm_cell_release_singleton with cell_defer_no_holder
In every case dm_cell_release_singleton is called the bio in question
is the holder.  If there are no non-holder entries in the cell then
cell_defer_no_holder (renamed from cell_defer_except) behaves
identically to dm_cell_release_singleton.  If there *are* non-holder
entries then dm_cell_release_singleton is the wrong thing to call
because those entries must be deferred.

This patch offers no functional change but is a pre-requisite for
"dm thin: eliminate race when io and discards are issued
simultaneously the same block".

Signed-off-by: Joe Thornber <ejt@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>
Cc: stable@vger.kernel.org