dm cache: avoid  the size of per_bio_data when using writeback
When wriethrough is used the per_bio_data structure includes a
dm_bio_details structure that is 1056 bytes.  This structure is rather
large so it is best to avoid allocating it for every bio if it can be
avoided.  If writeback is used we can avoid allocating the writethrough
specific memebers of the per_bio_data structure.

So per_bio_data_size is 16 bytes if writeback is used and 1096
bytes if writethrough is used.  In the near future, if/when Kent
Overstreet's bvec changes go upstream, the dm_bio_details structure will
be reduced to 32 bytes (nice improvement from it's current 1056 bytes)
-- leaving per_bio_data_size to be 72 bytes at that time.  So even
if/when that comes to pass there is still a benefit to conditionally
setting per_bio_data_size.

Signed-off-by: Mike Snitzer <snitzer@redhat.com>